fix: make SEO overrides work with server side rendering
