+++ web/package.json
@@ -103,7 +103,8 @@
     "reselect": "4.0.0",
     "serialize-javascript": "5.0.1",
     "styled-components": "5.1.1",
-    "typeface-open-sans": "0.0.75"
+    "typeface-open-sans": "0.0.75",
+    "url-join": "4.0.1"
   },
   "devDependencies": {
     "@babel/core": "7.11.5",
@@ -164,6 +165,7 @@
     "@types/serialize-javascript": "5.0.0",
     "@types/styled-components": "5.1.2",
     "@types/terser-webpack-plugin": "4.1.0",
+    "@types/url-join": "4.0.0",
     "@types/webpack": "4.41.22",
     "@types/webpackbar": "4.0.0",
     "@typescript-eslint/eslint-plugin": "4.0.1",
