+++ web/src/pages/_document.tsx
@@ -1,9 +1,15 @@
 import React from 'react'
 
-import NextDocument, { Html, Head, Main, NextScript, DocumentContext, DocumentInitialProps } from 'next/document'
+import { get } from 'lodash'
+import urljoin from 'url-join'
 import { ServerStyleSheet } from 'styled-components'
+import NextDocument, { Html, Head, Main, NextScript, DocumentContext, DocumentInitialProps } from 'next/document'
 
-import { PROJECT_NAME, DOMAIN, URL_MANIFEST_JSON, URL_FAVICON } from 'src/constants'
+import { DOMAIN, URL_MANIFEST_JSON, URL_FAVICON, TWITTER_USERNAME_FRIENDLY } from 'src/constants'
+import { SEO_DEFAULT, SEO_OVERRIDES } from 'src/seo/seoMetadata'
+
+const lang = 'en'
+const localeFull = 'en-US'
 
 export const GenericIcons = [16, 32, 96, 128, 196].map((size) => {
   const sizes = `${size}x${size}`
@@ -62,11 +68,18 @@ export default class Document extends NextDocument {
   }
 
   render() {
+    const pathname = this.props.dangerousAsPath
+    const url = urljoin(DOMAIN, pathname)
+
+    const { name, description, image } = {
+      ...SEO_DEFAULT,
+      ...get(SEO_OVERRIDES, pathname, SEO_DEFAULT),
+    }
+
     return (
-      <Html lang="en">
+      <Html lang={lang}>
         <Head>
           <meta charSet="UTF-8" />
-          <title>{PROJECT_NAME}</title>
 
           <meta name="theme-color" content="#ffffff" />
           <link rel="manifest" href={URL_MANIFEST_JSON} />
@@ -78,6 +91,32 @@ export default class Document extends NextDocument {
           <link rel="mask-icon" href={`${DOMAIN}/icons/safari-pinned-tab.svg" color="#555555`} />
 
           {MicrosoftIcons}
+
+          <title>{name}</title>
+
+          <meta name="description" content={description} />
+          <meta name="application-name" content={name} />
+
+          <meta key="description" itemProp="description" content={description} />
+          <meta key="image" itemProp="image" content={image} />
+          <meta key="name" itemProp="name" content={name} />
+          <meta key="og:description" property="og:description" content={description} />
+          <meta key="og:image" property="og:image" content={image} />
+          <meta key="og:image:secure_url" property="og:image:secure_url" content={image} />
+          <meta key="og:image:type" property="og:image:type" content="image/png" />
+          <meta key="og:image:width" property="og:image:width" content="1200" />
+          <meta key="og:image:height" property="og:image:height" content="600" />
+          <meta key="og:locale" property="og:locale" content={localeFull} />
+          <meta key="og:title" property="og:title" content={name} />
+          <meta key="og:type" property="og:type" content="website" />
+          <meta key="og:url" property="og:url" content={url} />
+          <meta name="twitter:card" content="summary_large_image" />
+          <meta name="twitter:description" content={description} />
+          <meta name="twitter:image" content={image} />
+          <meta name="twitter:image:alt" content={description} />
+          <meta name="twitter:title" content={name} />
+          <meta name="twitter:url" content={url} />
+          <meta name="twitter:site" content={TWITTER_USERNAME_FRIENDLY} />
         </Head>
 
         <body>
